test: skip SQL view delete for SQLite

+++ scripts/cypress/integration/common/4h_kanban.js
@@ -164,9 +164,12 @@ export const genTest = (apiType, dbType) => {
 
       if (dbType === "postgres" || dbType === "xcdb") {
         cy.openTableTab("Film", 25);
-        // delete SQL views
-        cy.deleteTable("NicerButSlowerFilmList");
-        cy.deleteTable("FilmList");
+
+        if (dbType === "postgres") {
+          // delete SQL views
+          cy.deleteTable("NicerButSlowerFilmList");
+          cy.deleteTable("FilmList");
+        }
 
         // edit `rating` column: from custom DB type to single select
         editColumn();
